+++ view/src/const.ts
@@ -1,1 +1,2 @@
 export const kCommunicationToken = 'communication-token'
+export const kTerminalSettings = 'terminal-settings'