+++ view/src/lib/route.ts
@@ -2,4 +2,6 @@ export const kHomeRoute = '/'
 
 export const kDashboardRoute = '/dashboard'
 
+export const kTerminalRoute = '/terminal'
+
 export const kSettingsRoute = '/settings'