+++ view/src/routes/root.tsx
@@ -7,7 +7,7 @@ import { TailwindIndicator } from '@/components/tailwind-indicator'
 export default function Root() {
     return (
         <>
-            <div className="relative flex h-full min-h-screen flex-col">
+            <div className="relative flex h-full flex-col">
                 <SiteHeader />
                 <div className="container h-full items-center pb-8 pt-4">
                     <Outlet />