+++ view/src/routes/router.tsx
@@ -6,6 +6,8 @@ import SettingsLayout from '@/routes/settings/layout.tsx'
 import GeneralPage from '@/routes/settings/page.tsx'
 import SettingsSecurityPage from '@/routes/settings/security/page.tsx'
 import SettingsServerPage from '@/routes/settings/server/page.tsx'
+import SettingsTerminalPage from '@/routes/settings/terminal/page.tsx'
+import TerminalPage from '@/routes/terminal'
 import { createBrowserRouter, Navigate } from 'react-router-dom'
 
 const router = createBrowserRouter([
@@ -27,6 +29,10 @@ const router = createBrowserRouter([
                 element: <LoginPage />,
             },
             {
+                path: '/terminal',
+                element: <TerminalPage />,
+            },
+            {
                 path: '/settings',
                 element: <SettingsLayout />,
                 children: [
@@ -45,6 +51,10 @@ const router = createBrowserRouter([
                     {
                         path: '/settings/server',
                         element: <SettingsServerPage />,
+                    },
+                    {
+                        path: '/settings/terminal',
+                        element: <SettingsTerminalPage />,
                     },
                 ],
             },