+++ view/src/routes/settings/components/sidebar-nav.tsx
@@ -25,6 +25,7 @@ export function SidebarNav({ className, items, ...props }: SidebarNavProps) {
         >
             {items.map((item) => (
                 <Link
+                    draggable="false"
                     key={item.href}
                     to={item.disabled ? '' : item.href}
                     className={cn(