+++ web/src/main.rs
@@ -11,6 +11,8 @@ use crate::report::reporter::Reporter;
 use crate::route::config_route::config_services;
 use crate::route::local_route::local_services;
 use crate::route::page_route::page_services;
+#[cfg(not(target_os = "windows"))]
+use crate::route::pty_route::pty_service;
 use crate::server::echo_ws;
 use actix_web::{middleware, web, App, HttpServer};
 use clap::Parser;
@@ -21,6 +23,8 @@ mod config;
 mod db;
 mod handler;
 mod model;
+
+mod pty;
 mod report;
 mod route;
 mod server;
@@ -28,6 +32,7 @@ mod system_info;
 mod test;
 mod token;
 mod traits;
+mod utils;
 mod vo;
 
 #[actix_web::main]
@@ -51,21 +56,27 @@ async fn main() -> std::io::Result<()> {
     });
 
     HttpServer::new(move || {
-        App::new()
+        let mut app = App::new()
             .app_data(web::Data::new(Arc::clone(&config)))
             .app_data(web::JsonConfig::default().limit(4096))
             .configure(config_services)
             .configure(|cfg| local_services(cfg, &host))
             .service(web::resource("/version").to(version))
-            // .service(web::resource("/config").to(config_test))
-            .service(web::resource("/check").route(web::post().to(check_token)))
+            .service(web::resource("/check").to(check_token))
             .service(kill_process)
             .service(rest_token)
             // websocket route
             .service(web::resource("/ws").route(web::get().to(echo_ws)))
             .configure(page_services)
             // enable logger
-            .wrap(middleware::Logger::default())
+            .wrap(middleware::Logger::default());
+
+        #[cfg(not(target_os = "windows"))]
+        {
+            app = app.configure(pty_service);
+        }
+
+        app
     })
     .workers(2)
     .bind(("0.0.0.0", port))?